+++ app/db/config.ts
@@ -62,6 +62,24 @@ export const updateConfig = async ({
   });
 };
 
+/**
+ * Remove config from the DB.
+ * Only valid for logged in users.
+ *
+ * @param key Key of the config to be updated
+ */
+export const removeConfig = async ({
+  key,
+}: {
+  key: string;
+}): Promise<{ key: string }> => {
+  return await prisma.config.delete({
+    where: {
+      key,
+    },
+  });
+};
+
 const migrateDataSet = (dataSet: string): string => {
   if (dataSet.includes("https://environment.ld.admin.ch/foen/nfi")) {
     return dataSet.replace(/None-None-/, "");
@@ -143,7 +161,11 @@ export const getUserConfigs = async (userId: number) => {
     where: {
       user_id: userId,
     },
+    orderBy: {
+      created_at: "desc",
+    },
   });
+
   return configs.map((c) => parseDbConfig(c));
 };