+++ app/login/components/profile-content-tabs.tsx
@@ -1,5 +1,6 @@
+import { t } from "@lingui/macro";
 import { TabContext, TabList, TabPanel } from "@mui/lab";
-import { Box, Tab, Theme, Typography } from "@mui/material";
+import { Box, Tab, Theme } from "@mui/material";
 import { makeStyles } from "@mui/styles";
 import clsx from "clsx";
 import React from "react";
@@ -35,11 +36,13 @@ const useStyles = makeStyles<Theme>((theme) => ({
 }));
 
 type ProfileContentTabsProps = {
+  userId: number;
   userConfigs: ParsedConfig[];
 };
 
 export const ProfileContentTabs = (props: ProfileContentTabsProps) => {
-  const { userConfigs } = props;
+  const { userId, userConfigs: _userConfigs } = props;
+  const [userConfigs, setUserConfigs] = React.useState(_userConfigs);
   const [value, setValue] = React.useState("Home");
   const handleChange = useEvent((_: React.SyntheticEvent, v: string) => {
     setValue(v);
@@ -53,34 +56,56 @@ export const ProfileContentTabs = (props: ProfileContentTabsProps) => {
         <Box className={rootClasses.sectionContent}>
           <Box sx={{ borderBottom: 1, borderColor: "divider" }}>
             <TabList className={classes.tabList} onChange={handleChange}>
-              {["Home", "My visualizations", "My favorite datasets"].map(
-                (d) => (
-                  <Tab key={d} className={classes.tab} label={d} value={d} />
-                )
-              )}
+              {[
+                "Home",
+                t({
+                  id: "login.profile.my-visualizations",
+                  message: "My visualizations",
+                }),
+              ].map((d) => (
+                <Tab key={d} className={classes.tab} label={d} value={d} />
+              ))}
             </TabList>
           </Box>
         </Box>
       </Box>
       <TabPanel className={classes.tabPanel} value="Home">
-        <Box className={classes.tabPanelContent}>
-          <ProfileVisualizationsTable userConfigs={userConfigs} />
+        <Box
+          className={classes.tabPanelContent}
+          sx={{ display: "flex", flexDirection: "column", gap: 6 }}
+        >
+          <ProfileVisualizationsTable
+            userId={userId}
+            userConfigs={userConfigs}
+            setUserConfigs={setUserConfigs}
+            preview
+            onShowAll={() =>
+              setValue(
+                t({
+                  id: "login.profile.my-visualizations",
+                  message: "My visualizations",
+                })
+              )
+            }
+          />
         </Box>
       </TabPanel>
-      <TabPanel className={classes.tabPanel} value="My visualizations">
+      <TabPanel
+        className={classes.tabPanel}
+        value={t({
+          id: "login.profile.my-visualizations",
+          message: "My visualizations",
+        })}
+      >
         <Box
           className={classes.tabPanelContent}
           sx={{ bgcolor: "background.paper" }}
         >
-          <ProfileVisualizationsTable userConfigs={userConfigs} />
-        </Box>
-      </TabPanel>
-      <TabPanel className={classes.tabPanel} value="My favorite datasets">
-        <Box
-          className={classes.tabPanelContent}
-          sx={{ bgcolor: "background.paper" }}
-        >
-          <Typography variant="h2">My favorite datasets</Typography>
+          <ProfileVisualizationsTable
+            userId={userId}
+            userConfigs={userConfigs}
+            setUserConfigs={setUserConfigs}
+          />
         </Box>
       </TabPanel>
     </TabContext>