+++ app/login/components/profile-header.tsx
@@ -1,8 +1,10 @@
+import { Trans } from "@lingui/macro";
 import { Box, Button, Link, Theme, Typography } from "@mui/material";
 import { makeStyles } from "@mui/styles";
 import { User } from "@prisma/client";
 import clsx from "clsx";
 import { signOut } from "next-auth/react";
+import NextLink from "next/link";
 
 import { useRootStyles } from "@/login/utils";
 
@@ -35,9 +37,7 @@ export const ProfileHeader = (props: ProfileHeaderProps) => {
       <Box className={rootClasses.sectionContent}>
         <Box className={classes.topRow}>
           <Typography variant="h1">{user.name}</Typography>
-          <Button component={Link} href="/browse">
-            Browse all datasets
-          </Button>
+          <BrowseButton />
         </Box>
         <Button
           className={classes.browseButton}
@@ -50,5 +50,23 @@ export const ProfileHeader = (props: ProfileHeaderProps) => {
         </Button>
       </Box>
     </Box>
+  );
+};
+
+const BrowseButton = () => {
+  return (
+    <Button>
+      <NextLink href="/browse" passHref legacyBehavior>
+        <Link
+          sx={{
+            "&:hover": {
+              textDecoration: "none",
+            },
+          }}
+        >
+          <Trans id="browse.dataset.all">Browse all datasets</Trans>
+        </Link>
+      </NextLink>
+    </Button>
   );
 };