+++ CHANGELOG.md
@@ -11,6 +11,8 @@ You can also check the [release page](https://github.com/visualize-admin/visuali
 
 - Features
   - Added a new group of charts – Combo charts – that includes multi-measure line, dual-axis line and column-line charts
+  - Added a way to edit and remove charts for logged in users
+  - Improved user profile page
 
 # [3.22.9] - 2023-10-06